Add retries to backwards incompat cache recreation
This moves a recently created retry function to a place where it can be
imported by both the adapter and the server, and then uses it when
trying to recreate caches after a backwards incompatible upgrade, so
that we have a lower chance of false negatives due to transient
failures.
